+++ agent-control/src/agent_control/run/k8s.rs
@@ -168,7 +168,8 @@ impl AgentControlRunner {
 
         let mut value_providers = ValueProviders::default()
             .with_env()
-            .with_k8s_secret(k8s_client.clone());
+            .with_k8s_secret(k8s_client.clone())
+            .with_k8s_configmap(k8s_client.clone());
         if let Some(config) = &agent_control_config.value_providers {
             value_providers = value_providers
                 .with_config(config.clone())